requirements.txt
certifi==2025.10.5
charset-normalizer==3.4.4
docker==7.1.0
fastapi==0.121.0
idna==3.11
psutil==7.1.3
pydantic==2.12.3
python-multipart==0.0.20
requests==2.32.5
sniffio==1.3.1
starlette==0.49.3
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
websockets==15.0.1
python-telegram-bot==20.0
python-dotenv==0.19.2
websocket-client